Updated commit following review.
