+++ cheddar/cheddar-persistence/src/main/java/com/clicktravel/cheddar/infrastructure/persistence/exception/PersistenceResourceFailureException.java
@@ -14,21 +14,19 @@
  * limitations under the License.
  * 
  */
-package com.clicktravel.cheddar.infrastructure.persistence.database.exception;
+package com.clicktravel.cheddar.infrastructure.persistence.exception;
 
-public abstract class PersistenceException extends RuntimeException {
+public class PersistenceResourceFailureException extends PersistenceException {
 
     private static final long serialVersionUID = 1L;
 
-    public PersistenceException(final String message) {
-        super(message);
-    }
-
     /**
-     * @param message
-     * @param cause
+     * Constructor for PersistenceResourceFailureException
+     * 
+     * @param message The message associated with the exception
+     * @param cause The cause of the exception
      */
-    public PersistenceException(final String message, final Throwable cause) {
+    public PersistenceResourceFailureException(final String message, final Throwable cause) {
         super(message, cause);
     }
